fixed rccl kernels to work with new hip
