+++ src/rcclScalarReduceKernels.h
@@ -17,7 +17,7 @@ All rights reserved.
 //! @brief Definition of RcclKernelScalarReduce
 //! Gather data from non-root gpus and do reduction op on it
 template <typename DataType_t, rcclRedOp_t Op>
-__global__ void RcclKernelScalarReduce(RingNode_t* pcurr_track, void* send_buff,
+__global__ void RcclKernelScalarReduce(RingNode_t* pcurr_track, const void* send_buff,
                                        void* recv_buff, int count) {
     int tx = threadIdx.x;
     int bx = blockIdx.x;
@@ -25,7 +25,7 @@ __global__ void RcclKernelScalarReduce(RingNode_t* pcurr_track, void* send_buff,
 
     //! Get pointers to current gpu source and destination buffers
     DataType_t* curr_dst_buff = reinterpret_cast<DataType_t*>(recv_buff);
-    DataType_t* curr_src_buff = reinterpret_cast<DataType_t*>(send_buff);
+    const DataType_t* curr_src_buff = reinterpret_cast<const DataType_t*>(send_buff);
 
     //! Use only count number of workitems to do the reduction operation
     if (tid < count) {
